requirements.txt
sliver-py
requests
asyncio